+++ nippo_kintai_project/src/main/java/com/analix/project/entity/DailyReportDetail.java
@@ -13,7 +13,7 @@ public class DailyReportDetail {
 	/*日付*/
 	private LocalDate date;
 	/*時間*/
-	private Float time;
+	private Integer time;
 	/*内容*/
 	private String content;
